Split the platform dependent bits of Hijrah date
